Moved SubjectValueCache to another package
Part of #588
