+++ src/main/java/com/sap/oss/phosphor/fosstars/data/SubjectValueCache.java
@@ -1,7 +1,5 @@
-package com.sap.oss.phosphor.fosstars.tool.github;
+package com.sap.oss.phosphor.fosstars.data;
 
-import com.sap.oss.phosphor.fosstars.data.StandardValueCache;
-import com.sap.oss.phosphor.fosstars.data.ValueCache;
 import com.sap.oss.phosphor.fosstars.model.Feature;
 import com.sap.oss.phosphor.fosstars.model.Subject;
 import com.sap.oss.phosphor.fosstars.model.Value;
