+++ src/test/java/com/sap/oss/phosphor/fosstars/data/github/HasSecurityPolicyTest.java
@@ -11,10 +11,10 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
 import com.sap.oss.phosphor.fosstars.data.NoValueCache;
+import com.sap.oss.phosphor.fosstars.data.SubjectValueCache;
 import com.sap.oss.phosphor.fosstars.model.Value;
 import com.sap.oss.phosphor.fosstars.model.subject.oss.GitHubProject;
 import com.sap.oss.phosphor.fosstars.model.value.ValueHashSet;
-import com.sap.oss.phosphor.fosstars.tool.github.SubjectValueCache;
 import java.io.IOException;
 import java.util.Optional;
 import org.apache.commons.lang3.StringUtils;
